fix: rename auth controller
